requirements.txt
alembic==1.7.5
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.13
Flask==2.0.2
Flask-Caching==1.10.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
Pygments==2.10.0
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
rich==10.9.0
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.6
Werkzeug==2.0.2
wrapt==1.13.3
